// Neptune.Web/src/app/shared/generated/model/treatment-bmp-assessment-photo-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPAssessmentPhotoDto { 
    TreatmentBMPAssessmentPhotoID?: number;
    TreatmentBMPAssessmentID?: number;
    FileResourceID?: number;
    FileResourceGUID?: string | null;
    Caption?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPAssessmentPhotoDtoForm { 
    TreatmentBMPAssessmentPhotoID?: FormControl<number>;
    TreatmentBMPAssessmentID?: FormControl<number>;
    FileResourceID?: FormControl<number>;
    FileResourceGUID?: FormControl<string>;
    Caption?: FormControl<string>;
}

export class TreatmentBMPAssessmentPhotoDtoFormControls { 
    public static TreatmentBMPAssessmentPhotoID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPAssessmentID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FileResourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FileResourceGUID = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Caption = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}