// Neptune.Web/src/app/shared/generated/model/field-visit-workflow-dto.ts
export class FieldVisitWorkflowDto { 
    FieldVisitID?: number;
    VisitDate?: string;
    FieldVisitTypeID?: number;
    FieldVisitTypeDisplayName?: string | null;
    FieldVisitStatusID?: number;
    FieldVisitStatusDisplayName?: string | null;
    IsFieldVisitVerified?: boolean;
    InventoryUpdated?: boolean;
    TreatmentBMPID?: number;
    TreatmentBMPName?: string | null;
    TreatmentBMPTypeID?: number;
    TreatmentBMPTypeName?: string | null;
    StormwaterJurisdictionID?: number;
    PerformedByPersonID?: number;
    PerformedByPersonName?: string | null;
    InitialAssessmentID?: number | null;
    InitialAssessmentComplete?: boolean;
    InitialAssessmentScore?: number | null;
    PostMaintenanceAssessmentID?: number | null;
    PostMaintenanceAssessmentComplete?: boolean;
    PostMaintenanceAssessmentScore?: number | null;
    MaintenanceRecordID?: number | null;
    NumberOfRequiredAttributes?: number;
    NumberRequiredAttributesEntered?: number;
    readonly RequiredAttributesEntered?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}